feat(cruds): telas do frontend(crud completo de crianças finalizado, crud de planos e atividades ainda em construção)

+++ projeto_aps_back/src/main/java/br/edu/ifce/projetoapsback/repository/UserRepository.java
@@ -3,9 +3,10 @@ package br.edu.ifce.projetoapsback.repository;
 import br.edu.ifce.projetoapsback.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByEmail(String email);
     Optional<User> findByProfessionalCode(String professionalCode);
